feat(base-activity): Split Base Activity to BaseActivity & BaseActivityVM
